Refactored, made sure ResultDTO used by rabbit is serializable, started on score

+++ result/src/main/java/com/quiz/result/model/Result.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class Result {
-    private int resultId;
+    private Long resultId;
     private final int correctAlternative;
     private final String explanation;
